2026/day-34/docker-compose-advanced/requirements.txt
flask
psycopg2-binary
redis